# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orbit API",
        "description": "API de autenticación con Laravel Sanctum",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obtener todas las tareas",
                "operationId": "b07f65b771b24b2f1a573bbaef391dad",
                "responses": {
                    "200": {
                        "description": "Lista de tareas"
                    },
                    "404": {
                        "description": "No se encontraron tareas"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/store": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "a6d15c3a88000556710be28e8d10acc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "estado",
                                    "prioridad"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Comprar alimentos"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Comprar frutas, verduras y leche"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "prioridad": {
                                        "type": "string",
                                        "example": "alta"
                                    },
                                    "fecha_limite": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "task_list_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada"
                    },
                    "422": {
                        "description": "Validacion fallida"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/update/{id}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualizar una tarea existente",
                "operationId": "12381a24436a0daf1edac8777fc5ce3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "estado",
                                    "prioridad"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Comprar alimentos"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Comprar frutas, verduras y leche"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "prioridad": {
                                        "type": "string",
                                        "example": "alta"
                                    },
                                    "fecha_limite": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "task_list_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/delete/{id}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "516948911632f814d15559b11348ea79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}